2001-08-23  Mark McLoughlin <mark@skynet.ie>
        Bunch of warnings fixes.

        * libIDL/lexer.l, libIDL/parser.y,
          libIDL/util.c,
          src/idl-compiler/orbit-idl-c-headers.c:
        cast char to int for isspace() etc.

        * src/orb/GIOP/giop-recv-buffer.c
        (giop_recv_list_push): kill G_BREAKPOINT.

        * src/orb/dynamic/dynany.c
        (DynamicAny_DynStruct_get_elements_as_dyn)
        kill unused newdyn.

        * src/orb/orb-core/corba-context.c
        (ctx_get_values) changes CORBA_Identifier arg
        to const CORBA_char *.

        * src/orb/orb-core/iop-profiles.c
        (IOP_components_marshal): kill unsed lenptr.

        * test/test-dynany.c(test_struct): pass dyn_struct
        to DynamicAny_DynStruct_* instead of dyn_any.

+++ src/idl-compiler/orbit-idl-c-headers.c
@@ -427,9 +427,9 @@ ch_output_codefrag(IDL_tree tree, OIDL_Run_Info *rinfo, OIDL_C_Info *ci)
 		sizeof("#pragma include_defs")-1)) {
 	char *ctmp, *cte;
 	ctmp = ((char *)list->data) + sizeof("#pragma include_defs");
-	while(*ctmp && (isspace(*ctmp) || *ctmp == '"')) ctmp++;
+	while(*ctmp && (isspace((int)*ctmp) || *ctmp == '"')) ctmp++;
 	cte = ctmp;
-	while(*cte && !isspace(*cte) && *cte != '"') cte++;
+	while(*cte && !isspace((int)*cte) && *cte != '"') cte++;
 	*cte = '\0';
       fprintf(ci->fh, "#include <%s>\n", ctmp);
     } else
